combine the two chunks of phone-related protocols together
