+++ go/libkb/notify_router.go
@@ -83,9 +83,9 @@ type NotifyListener interface {
 	WalletRequestStatusNotification(reqID stellar1.KeybaseRequestID)
 	TeamListUnverifiedChanged(teamName string)
 	CanUserPerformChanged(teamName string)
-	PhoneNumberAdded(phoneNumber string)
-	PhoneNumberVerified(phoneNumber string)
-	PhoneNumberSuperseded(phoneNumber string)
+	PhoneNumberAdded(phoneNumber keybase1.PhoneNumber)
+	PhoneNumberVerified(phoneNumber keybase1.PhoneNumber)
+	PhoneNumberSuperseded(phoneNumber keybase1.PhoneNumber)
 }
 
 type NoopNotifyListener struct{}
@@ -168,9 +168,9 @@ func (n *NoopNotifyListener) WalletPaymentStatusNotification(accountID stellar1.
 func (n *NoopNotifyListener) WalletRequestStatusNotification(reqID stellar1.KeybaseRequestID) {}
 func (n *NoopNotifyListener) TeamListUnverifiedChanged(teamName string)                       {}
 func (n *NoopNotifyListener) CanUserPerformChanged(teamName string)                           {}
-func (n *NoopNotifyListener) PhoneNumberAdded(phoneNumber string)                             {}
-func (n *NoopNotifyListener) PhoneNumberVerified(phoneNumber string)                          {}
-func (n *NoopNotifyListener) PhoneNumberSuperseded(phoneNumber string)                        {}
+func (n *NoopNotifyListener) PhoneNumberAdded(phoneNumber keybase1.PhoneNumber)               {}
+func (n *NoopNotifyListener) PhoneNumberVerified(phoneNumber keybase1.PhoneNumber)            {}
+func (n *NoopNotifyListener) PhoneNumberSuperseded(phoneNumber keybase1.PhoneNumber)          {}
 
 // NotifyRouter routes notifications to the various active RPC
 // connections. It's careful only to route to those who are interested
@@ -1691,7 +1691,7 @@ func (n *NotifyRouter) HandleAvatarUpdated(ctx context.Context, name string, for
 	}
 }
 
-func (n *NotifyRouter) HandlePhoneNumberAdded(ctx context.Context, phoneNumber string) {
+func (n *NotifyRouter) HandlePhoneNumberAdded(ctx context.Context, phoneNumber keybase1.PhoneNumber) {
 	if n == nil {
 		return
 	}
@@ -1710,7 +1710,7 @@ func (n *NotifyRouter) HandlePhoneNumberAdded(ctx context.Context, phoneNumber s
 	}
 }
 
-func (n *NotifyRouter) HandlePhoneNumberVerified(ctx context.Context, phoneNumber string) {
+func (n *NotifyRouter) HandlePhoneNumberVerified(ctx context.Context, phoneNumber keybase1.PhoneNumber) {
 	if n == nil {
 		return
 	}
@@ -1729,7 +1729,7 @@ func (n *NotifyRouter) HandlePhoneNumberVerified(ctx context.Context, phoneNumbe
 	}
 }
 
-func (n *NotifyRouter) HandlePhoneNumberSuperseded(ctx context.Context, phoneNumber string) {
+func (n *NotifyRouter) HandlePhoneNumberSuperseded(ctx context.Context, phoneNumber keybase1.PhoneNumber) {
 	if n == nil {
 		return
 	}
