+++ protocol/avdl/keybase1/phone_numbers.avdl
@@ -2,18 +2,34 @@
 protocol phoneNumbers {
     import idl "common.avdl";
 
+    /**
+     Phone number support for TOFU chats.
+    */
+    record UserPhoneNumber {
+      @jsonkey("phone_number")
+      PhoneNumber phoneNumber;
+      @jsonkey("verified")
+      boolean verified;
+      @jsonkey("ctime")
+      UnixTime ctime;
+    }
+
+    void addPhoneNumber(int sessionID, PhoneNumber phoneNumber);
+    void verifyPhoneNumber(int sessionID, PhoneNumber phoneNumber, string code);
+    array<UserPhoneNumber> getPhoneNumbers(int sessionID);
+
     record PhoneNumberAddedMsg {
         @jsonkey("phone")
-        string phoneNumber;
+        PhoneNumber phoneNumber;
     }
 
     record PhoneNumberVerifiedMsg {
         @jsonkey("phone")
-        string phoneNumber;
+        PhoneNumber phoneNumber;
     }
 
     record PhoneNumberSupersededMsg {
         @jsonkey("phone")
-        string phoneNumber;
+        PhoneNumber phoneNumber;
     }
 }
