+++ protocol/json/keybase1/phone_numbers.json
@@ -9,10 +9,32 @@
   "types": [
     {
       "type": "record",
+      "name": "UserPhoneNumber",
+      "fields": [
+        {
+          "type": "PhoneNumber",
+          "name": "phoneNumber",
+          "jsonkey": "phone_number"
+        },
+        {
+          "type": "boolean",
+          "name": "verified",
+          "jsonkey": "verified"
+        },
+        {
+          "type": "UnixTime",
+          "name": "ctime",
+          "jsonkey": "ctime"
+        }
+      ],
+      "doc": "Phone number support for TOFU chats."
+    },
+    {
+      "type": "record",
       "name": "PhoneNumberAddedMsg",
       "fields": [
         {
-          "type": "string",
+          "type": "PhoneNumber",
           "name": "phoneNumber",
           "jsonkey": "phone"
         }
@@ -23,7 +45,7 @@
       "name": "PhoneNumberVerifiedMsg",
       "fields": [
         {
-          "type": "string",
+          "type": "PhoneNumber",
           "name": "phoneNumber",
           "jsonkey": "phone"
         }
@@ -34,13 +56,56 @@
       "name": "PhoneNumberSupersededMsg",
       "fields": [
         {
-          "type": "string",
+          "type": "PhoneNumber",
           "name": "phoneNumber",
           "jsonkey": "phone"
         }
       ]
     }
   ],
-  "messages": {},
+  "messages": {
+    "addPhoneNumber": {
+      "request": [
+        {
+          "name": "sessionID",
+          "type": "int"
+        },
+        {
+          "name": "phoneNumber",
+          "type": "PhoneNumber"
+        }
+      ],
+      "response": null
+    },
+    "verifyPhoneNumber": {
+      "request": [
+        {
+          "name": "sessionID",
+          "type": "int"
+        },
+        {
+          "name": "phoneNumber",
+          "type": "PhoneNumber"
+        },
+        {
+          "name": "code",
+          "type": "string"
+        }
+      ],
+      "response": null
+    },
+    "getPhoneNumbers": {
+      "request": [
+        {
+          "name": "sessionID",
+          "type": "int"
+        }
+      ],
+      "response": {
+        "type": "array",
+        "items": "UserPhoneNumber"
+      }
+    }
+  },
   "namespace": "keybase.1"
 }
