+++ protocol/json/keybase1/user.json
@@ -276,34 +276,6 @@
           "name": "res"
         }
       ]
-    },
-    {
-      "type": "record",
-      "name": "PhoneNumber",
-      "fields": [],
-      "doc": "Phone number support for TOFU chats.",
-      "typedef": "string"
-    },
-    {
-      "type": "record",
-      "name": "UserPhoneNumber",
-      "fields": [
-        {
-          "type": "PhoneNumber",
-          "name": "phoneNumber",
-          "jsonkey": "phone_number"
-        },
-        {
-          "type": "boolean",
-          "name": "verified",
-          "jsonkey": "verified"
-        },
-        {
-          "type": "UnixTime",
-          "name": "ctime",
-          "jsonkey": "ctime"
-        }
-      ]
     }
   ],
   "messages": {
@@ -689,48 +661,6 @@
       ],
       "response": "NextMerkleRootRes",
       "doc": "FindNextMerkleRootAfterReset finds the first Merkle root that contains the UID reset\n   at resetSeqno. You should pass it prev, which was the last known Merkle root at the time of\n   the reset. Usually, we'll just turn up the next Merkle root, but not always."
-    },
-    "addPhoneNumber": {
-      "request": [
-        {
-          "name": "sessionID",
-          "type": "int"
-        },
-        {
-          "name": "phoneNumber",
-          "type": "PhoneNumber"
-        }
-      ],
-      "response": null
-    },
-    "verifyPhoneNumber": {
-      "request": [
-        {
-          "name": "sessionID",
-          "type": "int"
-        },
-        {
-          "name": "phoneNumber",
-          "type": "PhoneNumber"
-        },
-        {
-          "name": "code",
-          "type": "string"
-        }
-      ],
-      "response": null
-    },
-    "getPhoneNumbers": {
-      "request": [
-        {
-          "name": "sessionID",
-          "type": "int"
-        }
-      ],
-      "response": {
-        "type": "array",
-        "items": "UserPhoneNumber"
-      }
     }
   },
   "namespace": "keybase.1"
